Create project tables via database migration runner (#12)
Added deployment database migration runner and table creation sql.
Important to note that migration script is not used in docker development, rather the docker conventional startup script is used. `migrations` table will be empty on local if using docker environment (recommended).

PR builds run the migration scripts, but on an empty DB so any data conversion issues if columns are being altered will not be caught until deployment failure. This should be addressed if it becomes a problem but for our purposes it should be ok.

Creates all tables IAW design documents in report 1 & 2 with minor fixes noticed during development (datatypes mostly).

Co-authored-by: jrich8573 <jasonrich85@icloud.com>
Co-authored-by: Suraksha <stasg001@odu.edu>
Co-authored-by: Jason Rich <jrich069@odu.edu>