Refactor codebase into M(V)C with DI (#20)
* Add php-di and fast-route as deps
* Major refactor oh my god

When you code at 2 am you just keep going. This is what happens.
1. Change web server from nginx to apache to match cs.odu.edu
2. Use mod_rewrite on server to enable server side routing
3. Wire up dependency injection using php-di
4. Refactor project to better resemble an MVC architecture

This is a lot of code change, but a fair amount is just moving stuff
around. The biggest (and honestly most complicated) piece is the DI,
but when testing code that relies on a DB we'll be glad it's there.